// frontend/src/proto/driver.ts
export interface RouteStop {
    /**
     * @generated from protobuf field: int32 sequence = 1
     */
    sequence: number;
    /**
     * @generated from protobuf field: string area_id = 2
     */
    areaId: string;
    /**
     * @generated from protobuf field: bool is_station = 3
     */
    isStation: boolean;
    /**
     * @generated from protobuf field: int32 arrival_offset_minutes = 4
     */
    arrivalOffsetMinutes: number;
}
// ...
class RouteStop$Type extends MessageType<RouteStop> {
    constructor() {
        super("lastmile.driver.RouteStop", [
            { no: 1, name: "sequence", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 2, name: "area_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "is_station", kind: "scalar", T: 8 /*ScalarType.BOOL*/ },
            { no: 4, name: "arrival_offset_minutes", kind: "scalar", T: 5 /*ScalarType.INT32*/ }
        ]);
    }
    create(value?: PartialMessage<RouteStop>): RouteStop {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.sequence = 0;
        message.areaId = "";
        message.isStation = false;
        message.arrivalOffsetMinutes = 0;
        if (value !== undefined)
            reflectionMergePartial<RouteStop>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: RouteStop): RouteStop {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* int32 sequence */ 1:
                    message.sequence = reader.int32();
                    break;
                case /* string area_id */ 2:
                    message.areaId = reader.string();
                    break;
                case /* bool is_station */ 3:
                    message.isStation = reader.bool();
                    break;
                case /* int32 arrival_offset_minutes */ 4:
                    message.arrivalOffsetMinutes = reader.int32();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: RouteStop, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* int32 sequence = 1; */
        if (message.sequence !== 0)
            writer.tag(1, WireType.Varint).int32(message.sequence);
        /* string area_id = 2; */
        if (message.areaId !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.areaId);
        /* bool is_station = 3; */
        if (message.isStation !== false)
            writer.tag(3, WireType.Varint).bool(message.isStation);
        /* int32 arrival_offset_minutes = 4; */
        if (message.arrivalOffsetMinutes !== 0)
            writer.tag(4, WireType.Varint).int32(message.arrivalOffsetMinutes);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message lastmile.driver.RouteStop
 */
export const RouteStop = new RouteStop$Type();